ScrollWrapper.tsx added to App.tsx to ensure that new pages always open at the top.
App.tsx has a custom hook
